backend/requirements.txt
Flask==2.2.5
Flask-CORS==4.0.0
Werkzeug==2.2.3
Jinja2==3.1.2
PyPDF2==3.0.1
pdfplumber==0.10.3
Pillow==10.0.1
pytesseract==0.3.10
reportlab==4.0.4
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
pytest==7.4.2
pytest-flask==1.2.0 